+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.1.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-sync-for-commerce-version-1/0.1.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-09-12 00:15:12
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.1] previous-versions/sync-for-commerce-version-1
+### Releases
+- [PyPI v0.1.1] https://pypi.org/project/codat-sync-for-commerce-version-1/0.1.1 - previous-versions/sync-for-commerce-version-1